requirements.txt
Brotli==1.1.0
inflate64==0.3.1
multivolumefile==0.2.3
psutil==5.9.6
py7zr==0.20.6
pybcj==1.0.1
pycryptodomex==3.19.0
pyppmd==1.0.0
PyYAML==6.0.1
pyzstd==0.15.9
texttable==1.7.0
urllib3==2.0.7
zipfile36==0.1.3
